Improve mobiles store config
