+++ packages/mobile/src/message/handlers/download.ts
@@ -7,9 +7,9 @@ import type {
 } from 'rn-fetch-blob'
 import RNFetchBlob from 'rn-fetch-blob'
 
-import { dispatch } from 'app/App'
 import type { MessageHandlers } from 'app/message/types'
 import { MessageType } from 'app/message/types'
+import { dispatch } from 'app/store'
 import {
   setDownloadedPercentage,
   setFileInfo,
